requirements.txt
pillow
pyserial
pywin32
pywin32-ctypes
PySide6
pycaw
psutil
keyboard